tests: fix default namespace, and allow override
Always use the namespaces suggested in the official documentation.

Allow to override the namespace name the tests wants to consume;
this is needed because when doing validation, we need to make sure to
check what was actually installed in the cluster, and there is no hard
requirement on the namespace name - only suggestions.

Signed-off-by: Francesco Romani <fromani@redhat.com>